+++ src/ui_loader/ui_builder_default.c
@@ -90,8 +90,6 @@ static ret_t ui_builder_default_on_end(ui_builder_t* b) {
     if (widget->vt->is_window) {
       event_t e = event_init(EVT_WINDOW_LOAD, widget);
       widget_dispatch_recursive(widget, &e);
-
-      widget_layout(widget);
     }
   }